5/10/20
Mostly created a confusion matrix from Ytest and Ytrain and created subcategories from there. Changed the length of Xtrain,Xtest,Ytrain,Ytest to be 50/50 instead of 80/20 just because they have to be the same length for the confusion matrix. Also changed simulation number to 200 insteaf of 100 and N to 100 insteaf of 50.